==> analysis/5_report.rmd <==
# Gaza Vaccine Preventable Disease Immunity Modelling
```{r setup, echo=FALSE}
true_wd <- getwd()
split_wd <- str_split(true_wd, "/")[[1]]
if(tail(split_wd, 1) == "analysis") {
    true_wd <-
        paste0(
            split_wd[-length(split_wd)],
            collapse = "/"
        )
}
readRDS_2 <- function(file) {
    readRDS(file.path(true_wd, file))
}
read_csv_2 <- function(file, ...) {
    read_csv(file.path(true_wd, file), show_col_types = FALSE, ...)
}
```
This report summarises modelling to track the amount of immunity to various vaccine-preventabel diseases in Gaza over the course of the conflict.
The diseases currently modelled are:

- MMR:
    - Measles
- Pentavalent:
    - Diphtheria
    - Pertussis
    - Hib Disease
- Rotavirus
- PCV:
    - Pneumococcal Disease
- Polio:
    - Polio (Wild-type)
    - Polio (Vaccine-derived)

## Model Details

INSERT MODEL STRUCTURE DIAGRAM

The model is an ODE model that tacks the waning of immunity to vaccinations based on defined coverage targets.
The model also simulates aquired immunity from pre-defined force of infection (FOI) values.

### Key Assumptions

- Death rates within each age group are constant across vaccination and disease status. The relative rate of death in each age group is assumed to be follow the distribution reported from 2019-2022.
- Crude birth rates are used so the birth are relative to the whole population, proportion of those born with maternal immunity is based on the proportions immune in the population between 15-45.
- Vaccine coverages are modelled as the proportion of people ageing into the relevant age group, will revisit this when modelling vaccinations in crisis
- Disease incidence rates are reweighted to account for population immunity (i.e. risk in S is greater than the crude rate but capped at 1), give equation!
- To translate rates across age groups it is assumed that the rates are uniform within age groups, and to translate populations it is assumed the age distribution within each age group is flat.
- The model assumes that the population is closed and that there is no migration in or out of the population.
- The model assumes that the population is well-mixed and that there is no spatial structure to the population.
- VE is assumed to be the same across all doses, could apply some scaling.
- We model no vaccine coverage at the start of the model.

Current issues:

- This model is inappropriate for modelling Pneumococcal Disease 
- HiB Disease data is only Meningitis, and Pneumococcal Disease data is just any bacterial Meningitis
- No Pertussis data, though there is transmission in Lebanon
- Model currently only uses central estimates

## Data

### Birth Rate
```{r crude_birth_rate, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/crude_birth_rate.rds") %>%
    mutate(
        `Crude Birth Rate` = `Crude Birth Rate`*1000*365
    )
df %>%
    ggplot(aes(x = Date, y = `Crude Birth Rate`)) +
    geom_line() +
    ggpubr::theme_pubclean() +
    scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
    coord_cartesian(ylim = c(0, max(df$`Crude Birth Rate`))) +
    geom_vline(xintercept = date_projection_start, linetype = "dashed") +
    labs(x = "Date", y = "Yearly Crude Birth Rate\n(per 1000)")

```

The crude birth rate is taken from the CIA factbook, assumed to be constant with a year.

### Death Rate
```{r crude_death_rate, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/crude_death_rates.rds") %>%
    mutate(
        `Crude Death Rate` = `Crude Death Rate`*1000*365
    )
df %>%
    ggplot(aes(x = Date, y = `Crude Death Rate`)) +
    geom_line() +
    ggpubr::theme_pubclean() +
    scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
    coord_cartesian(ylim = c(0, max(df$`Crude Death Rate`))) +
    geom_vline(xintercept = date_projection_start, linetype = "dashed") +
    labs(x = "Date", y = "Yearly Crude Death Rate\n(per 1000)")
```

The crude death rate is taken from the World Bank, assumed to be constant with a year.

```{r death_rate_age, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/age_proportional_death_rate.rds")
df %>%
    ggplot(aes(x = age_group_start, xend = age_group_end, y = proportional_death_rate, yend = proportional_death_rate)) +
    geom_segment() +
    ggpubr::theme_pubclean() +
    labs(x = "Age", y = "Proportional risk of death compared to the other age groups")
```

As it is unreasonable to assume that e crude death rate applies flatly across all age groups, the inidividual instantaneous death rate $D_{t, a}$ is calculated by:
$$D_{t, a} = D_{t} \times \frac{P_{a} \times N_{t, a}}{\sum_{a} P_{a} \times N_{t, a}}$$
where $D_{t}$ is the crude death rate at time $t$, $P_{a}$ is the proportional rate of death in age group $a$ as shown above, and $N_{t, a}$ is the number of people in age group $a$ at time $t$.

### Force of Infection
```{r foi, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/foi.rds") %>%
    mutate(
        `Force of Infection` = `Force of Infection`*1000*365
    ) %>%
    group_by(Pathogen) %>%
    filter(any(`Force of Infection` > 0))
df %>%
    ggplot(aes(x = Date, y = `Force of Infection`, colour = Pathogen)) +
    geom_line() +
    ggpubr::theme_pubclean() +
    scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
    coord_cartesian(ylim = c(0, max(df$`Force of Infection`))) +
    geom_vline(xintercept = date_projection_start, linetype = "dashed") +
    labs(x = "Date", y = "Yearly Force of Infection\n(per 1000)")
```

Force of infection is taken from the yearly incidence rates from MoH Gaza, these are scaled to the force of infection on susceptible individual through the equation:
$$\lambda_{S, t} = min\left(\lambda{t} \times \frac{N_t}{S_t}, 1\right).$$

### Vaccination Coverage
```{r vaccination_coverage, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/vaccine_coverage.rds")
df %>%
    ggplot(aes(x = Date, y = Coverage, colour = Vaccine)) +
    geom_line() +
    ggpubr::theme_pubclean() +
    scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
    scale_y_continuous(labels = scales::percent) +
    coord_cartesian(ylim = c(0.75, 1)) +
    geom_vline(xintercept = date_projection_start, linetype = "dashed") +
    labs(x = "Date", y = "Vaccine coverage\n(assumed to apply to all relevant scheduled doses)")
```

### Initial Population
The initial population age breakdown is taken from WPP estimate for 2000 for Palestine as a whole and then scaled to match the PCBS population estimate of Gaza for 2000.

## Approach

Our current approach is use central estimates to produce a single run.
We as vaccination and immunity has no impact on demographics we can sense check the model against the population data.

### Scenario Assumptions
none atm

## Outputs

### Checking Demographics data

```{r total_age, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/demographics.rds")
#total population comparison to MoH and PCBS
read_csv_2("data/raw/total_pop_over_time.csv", skip = 1) %>%
    filter(Year == "Gaza Strip") %>%
    select(-Year) %>%
    pivot_longer(cols = everything(), names_to = "Year", values_to = "Population") %>%
    mutate(scenario = "MoH") %>%
    rbind(
        read_csv_2("data/raw/total_pop_over_time_PCBS.csv", skip = 3) %>%
            filter(Governorate == "Gaza Strip") %>%
            select(-Governorate) %>%
            pivot_longer(cols = everything(), names_to = "Year", values_to = "Population") %>%
            mutate(scenario = "PCBS")
    ) %>%
    mutate(Year = as.integer(Year)) %>%
    rbind(
        df %>%
            group_by(scenario, date) %>%
            summarise(Population = sum(population)) %>%
            mutate(Year = year(date)) %>%
            group_by(scenario, Year) %>%
            summarise(Population = mean(Population))
    ) %>%
    mutate(Year = ymd(paste0(Year, "-06-01"))) %>%
    ggplot(aes(x = Year, y = Population, colour = scenario)) +
        geom_line(linewidth = 1) +
        ggpubr::theme_pubclean() +
        geom_vline(xintercept = date_projection_start, linetype = "dashed") +
        scale_x_date(date_breaks = "1 year", date_labels = "%Y") +
        coord_cartesian(xlim = ymd(paste0(c(2000, 2024), "-06-01"))) +
        labs(x = "Date", y = "Population", colour = "Source:", title = "Population over the years")
```

```{r 2023_ages, echo=FALSE, out.width="100%"}
#age break down 2023
read_csv_2("data/raw/pop_2023_UNFPA.csv", skip = 2) %>%
    transmute(
        age_group = `...11`,
        population = `Total...14`
    ) %>%
    mutate(
        age_group = case_when(
            age_group == "0" ~ "0mo",
            age_group == "1- 4" ~ "1to59mo",
            age_group == "5 - 9" ~ "5to9yo",
            age_group == "10 - 14" ~ "10to14yo",
            age_group == "15 - 19" ~ "15to19yo",
            age_group %in% c("20 - 24", "25 - 29") ~ "20to29yo",
            age_group %in% c("30 - 34", "35 - 39") ~ "30to39yo",
            age_group %in% c("40 - 44", "45 - 49") ~ "40to49yo",
            age_group %in% c("50 - 54", "55 - 59") ~ "50to59yo",
            age_group %in% c("60 - 64", "65 - 69", "70 - 74", "75 - 79", "80 +") ~ "60to100yo",
            TRUE ~ NA_character_
        )
    ) %>%
    group_by(age_group) %>%
    summarise(population = sum(population)) %>%
    filter(!is.na(age_group)) %>%
    mutate(source = "UNFPA") %>%
    rbind(
        df %>%
            filter(date >= ymd("2023-01-01") & scenario == "central") %>%
            group_by(age_group) %>%
            summarise(population = mean(population), .groups = "drop") %>%
            mutate(
                age_group = if_else(age_group %in% c("1to11mo", "12to59mo"), "1to59mo", age_group)
            ) %>%
            group_by(age_group) %>%
            summarise(population = sum(population), .groups = "drop") %>%
            mutate(source = "Model")
    ) %>%
    mutate(
        age_group = factor(age_group, levels = c("0mo", "1to59mo", "5to9yo", "10to14yo", "15to19yo", "20to29yo", "30to39yo", "40to49yo", "50to59yo", "60to100yo"))
    ) %>%
    ggplot(aes(x = age_group, y = population, fill = source)) +
        geom_col(position = "dodge") +
        ggpubr::theme_pubclean() +
        labs(x = "Age Group", y = "Population", fill = "Source:", title = "2023 Population Breakdown")
```

### Immunity projections

```{r immunity, echo=FALSE, out.width="100%"}
df <- readRDS_2("data/derived/projections_full.rds") %>%
    group_by(date, vaccine_type, scenario) %>%
    summarise(
        immunity = sum(immune)/sum(population),
        .groups = "drop"
    )

df %>% 
    ggplot(aes(x = date, y = immunity, colour = vaccine_type)) +
    geom_line() +
    ggpubr::theme_pubclean() +
    scale_x_date(date_breaks = "months", date_labels = "%b %Y") +
    scale_y_continuous(labels = scales::percent) +
    coord_cartesian(ylim = c(0, max(df$immunity))) +
    labs(x = "Date", y = "Proportion of Population Immune (%)", colour = "Disease:") +
    facet_wrap(~scenario, nrow = 1)
```
